+++ internal/testfixture/fakedmi/fakedmi_test.go
@@ -3,5 +3,5 @@ package fakedmi
 import "testing"
 
 func TestNew(t *testing.T) {
-	_ = New(t, nil)
+	_ = New(t)
 }